requirements.txt
streamlit
audio_recorder_streamlit
treamlit_float
os
openai
base64
dotenv